refactor: move explorer factories under deps
