+++ src/pages/colors/colors.handlers.js
@@ -1,6 +1,6 @@
 import { nanoid } from "nanoid";
 import { recursivelyCheckResource } from "../../utils/resourceUsageChecker.js";
-import { createResourceFileExplorerHandlers } from "../shared/fileExplorerHandlers.js";
+import { createResourceFileExplorerHandlers } from "../../deps/features/fileExplorerHandlers.js";
 
 const getColorItemById = ({ store, itemId } = {}) => {
   if (!itemId) return undefined;
