+++ src/pages/typography/typography.handlers.js
@@ -6,7 +6,7 @@ import {
 } from "../../constants/typography.js";
 import { getFileType } from "../../utils/fileTypeUtils";
 import { recursivelyCheckResource } from "../../utils/resourceUsageChecker.js";
-import { createResourceFileExplorerHandlers } from "../shared/fileExplorerHandlers.js";
+import { createResourceFileExplorerHandlers } from "../../deps/features/fileExplorerHandlers.js";
 
 // Helper function to sync repository state to store
 const syncRepositoryToStore = (store, projectService) => {
